Refactor `CommonModel`
`CommonModel` -> `CommonModelManager`
`Model` -> `CommonModel`
