+++ src/main/java/spleetwaise/commons/logic/commands/CommandUtil.java
@@ -8,7 +8,7 @@ import spleetwaise.address.logic.Messages;
 import spleetwaise.address.model.person.Person;
 import spleetwaise.commons.core.index.Index;
 import spleetwaise.commons.logic.commands.exceptions.CommandException;
-import spleetwaise.commons.model.CommonModel;
+import spleetwaise.commons.model.CommonModelManager;
 
 /**
  * Utility class for common command-related operations. Contains helper methods to reduce code duplication in command
@@ -19,12 +19,12 @@ public class CommandUtil {
     /**
      * Retrieves a {@code Person} from the filtered person list at the specified {@code Index}.
      *
-     * @param model The {@code CommonModel} instance used to retrieve the filtered person list.
+     * @param model The {@code CommonModelManager} instance used to retrieve the filtered person list.
      * @param index The {@code Index} of the person to retrieve from the filtered address book list.
      * @return The {@code Person} at the specified index in the filtered person list.
      * @throws CommandException If the specified index is out of bounds for the filtered person list.
      */
-    public static Person getPersonByFilteredPersonListIndex(CommonModel model, Index index)
+    public static Person getPersonByFilteredPersonListIndex(CommonModelManager model, Index index)
             throws CommandException {
         requireAllNonNull(model, index);
         List<Person> lastShownList = model.getFilteredPersonList();
