+++ src/test/java/spleetwaise/transaction/logic/commands/ListCommandTest.java
@@ -9,7 +9,7 @@ import spleetwaise.transaction.testutil.TypicalIndexes;
 import spleetwaise.transaction.testutil.TypicalTransactions;
 
 /**
- * Contains integration tests (interaction with the Model) and unit tests for ListCommand.
+ * Contains integration tests (interaction with the CommonModel) and unit tests for ListCommand.
  */
 public class ListCommandTest {
