+++ src/test/java/spleetwaise/transaction/logic/parser/FilterCommandParserTest.java
@@ -16,7 +16,7 @@ import spleetwaise.address.model.AddressBookModel;
 import spleetwaise.address.model.AddressBookModelManager;
 import spleetwaise.address.model.person.Person;
 import spleetwaise.address.testutil.TypicalPersons;
-import spleetwaise.commons.model.CommonModel;
+import spleetwaise.commons.model.CommonModelManager;
 import spleetwaise.transaction.logic.commands.FilterCommand;
 import spleetwaise.transaction.model.FilterCommandPredicate;
 import spleetwaise.transaction.model.TransactionBookModel;
@@ -57,7 +57,7 @@ public class FilterCommandParserTest {
     public static void setUp() {
         abModel.setAddressBook(TypicalPersons.getTypicalAddressBook());
         txnModel.setTransactionBook(TypicalTransactions.getTypicalTransactionBook());
-        CommonModel.initialise(abModel, txnModel);
+        CommonModelManager.initialise(abModel, txnModel);
     }
 
     @Test
