+++ src/test/java/spleetwaise/transaction/logic/parser/TransactionParserTest.java
@@ -12,7 +12,7 @@ import spleetwaise.address.model.AddressBookModelManager;
 import spleetwaise.address.testutil.TypicalPersons;
 import spleetwaise.commons.logic.commands.Command;
 import spleetwaise.commons.logic.parser.exceptions.ParseException;
-import spleetwaise.commons.model.CommonModel;
+import spleetwaise.commons.model.CommonModelManager;
 import spleetwaise.transaction.logic.commands.AddCommand;
 import spleetwaise.transaction.logic.commands.ClearCommand;
 import spleetwaise.transaction.logic.commands.EditCommand;
@@ -33,14 +33,14 @@ public class TransactionParserTest {
 
     @BeforeEach
     void setUp() {
-        CommonModel.initialise(abModel, tbModel);
+        CommonModelManager.initialise(abModel, tbModel);
     }
 
     @Test
     public void parseCommand_add() throws Exception {
         AddressBookModelManager aBModel = new AddressBookModelManager();
         aBModel.addPerson(TypicalPersons.ALICE);
-        CommonModel.initialise(aBModel, null);
+        CommonModelManager.initialise(aBModel, null);
         AddCommand command = (AddCommand) parser.parseCommand(TransactionUtil.getAddCommand());
         assertNotNull(command);
     }
